+++ pettingzoo/atari/basketball_pong_v3.py
@@ -1,5 +1,1 @@
-from .basketball_pong.basketball_pong import (  # noqa: F401
-    env,
-    parallel_env,
-    raw_env,
-)
+from .basketball_pong.basketball_pong import env, parallel_env, raw_env  # noqa: F401
