+++ pettingzoo/atari/combat_tank_v2.py
@@ -1,1 +1,1 @@
-from .combat_tank.combat_tank import env, raw_env, parallel_env  # noqa: F401
+from .combat_tank.combat_tank import env, parallel_env, raw_env  # noqa: F401
