+++ pettingzoo/atari/double_dunk_v3.py
@@ -1,1 +1,1 @@
-from .double_dunk.double_dunk import env, raw_env, parallel_env  # noqa: F401
+from .double_dunk.double_dunk import env, parallel_env, raw_env  # noqa: F401
