+++ pettingzoo/atari/entombed_competitive_v3.py
@@ -1,1 +1,5 @@
-from .entombed_competitive.entombed_competitive import env, raw_env, parallel_env  # noqa: F401
+from .entombed_competitive.entombed_competitive import (  # noqa: F401
+    env,
+    parallel_env,
+    raw_env,
+)
