+++ pettingzoo/atari/entombed_cooperative_v3.py
@@ -1,1 +1,1 @@
-from .entombed_cooperative.entombed_cooperative import env, raw_env, parallel_env  # noqa: F401
+from .entombed_cooperative.entombed_cooperative import env, parallel_env, raw_env  # noqa: F401
