+++ pettingzoo/atari/flag_capture_v2.py
@@ -1,1 +1,1 @@
-from .flag_capture.flag_capture import env, raw_env, parallel_env  # noqa: F401
+from .flag_capture.flag_capture import env, parallel_env, raw_env  # noqa: F401
