+++ pettingzoo/atari/foozpong_v3.py
@@ -1,1 +1,1 @@
-from .foozpong.foozpong import env, raw_env, parallel_env  # noqa: F401
+from .foozpong.foozpong import env, parallel_env, raw_env  # noqa: F401
