+++ pettingzoo/atari/joust_v3.py
@@ -1,1 +1,1 @@
-from .joust.joust import env, raw_env, parallel_env  # noqa: F401
+from .joust.joust import env, parallel_env, raw_env  # noqa: F401
