+++ pettingzoo/atari/mario_bros_v3.py
@@ -1,1 +1,1 @@
-from .mario_bros.mario_bros import env, raw_env, parallel_env  # noqa: F401
+from .mario_bros.mario_bros import env, parallel_env, raw_env  # noqa: F401
