+++ pettingzoo/atari/video_checkers_v4.py
@@ -1,17 +1,1 @@
-import os
-
-from .base_atari_env import BaseAtariEnv, base_env_wrapper_fn, parallel_wrapper_fn
-
-
-def raw_env(**kwargs):
-    return BaseAtariEnv(
-        game="video_checkers",
-        num_players=2,
-        mode_num=None,
-        env_name=os.path.basename(__file__)[:-3],
-        **kwargs
-    )
-
-
-env = base_env_wrapper_fn(raw_env)
-parallel_env = parallel_wrapper_fn(env)
+from .video_checkers.video_checkers import env, parallel_env, raw_env  # noqa: F401
